restructured src files and revised readme.md
